Перенес логику рендера board из main.js в контроллер
