+++ src/components/load-more-button/load-more-button.js
@@ -9,4 +9,9 @@ export default class LoadMoreButton extends AbstractComponent {
   getTemplate() {
     return createLoadMoreButtonTemplate();
   }
+
+  setClickHandler(handler) {
+    this.getElement()
+      .addEventListener(`click`, handler);
+  }
 }
