+++ src/components/task/task.js
@@ -93,4 +93,9 @@ export default class Task extends AbstractComponent {
   getTemplate() {
     return createTaskTemplate(this._task);
   }
+
+  setEditButtonClickHandler(handler) {
+    this.getElement().querySelector(`.card__btn--edit`)
+      .addEventListener(`click`, handler);
+  }
 }
